[refactor] Parser simplification refactor
his introduces a number of changes to simplify the generated Pparser structure. Firstly, we unify all individual characters into a single new class, the fake character class, that all individual characters inherit from. Next, we eliminate the recursive structure of the single-quote-literal and the  block-comment. While previously both had the form continue ::= TERMINAL continue | ENDING, we now mark the continue rule as private so that these are represented as a list and not a nested tree. Finally, several of the label nonterminals and nonterminals whose function is to represent character classes are eliminated from the produced tree structure for being unnecessary.