[fix] Fix parsing of block comments
Block comment parsing should not backtrack. The default, however, is for a subrule of block-comment-continue to fail to parse the nested block comment subrule if parsing a block-comment fails, and so falls through to the final block-comment-char case. This is the incorrect behavior for strings such as '{-{--}', which should be read as an incomplete block comment rather than a complete block comment containing the middle '{-' characters. To fix this, we use pinning rules to ensure that once a block comment is recognized by its left brace and minus we commit to parsing what follows as part of that block comment.