[feature] Add further highlights and initial highlighting test mechanism
The primary challenge in this round of highlighting was recognizing that when a child element being annotated is an error element, the default annotator behavior is to disable highlighting of parent elements. This would requiree highlighting to be placed as closely to the leaf elements as possible, which seems inefficient because  each character has its own leaf node. Instead, we use the highlight range extension point to force parents to be highlighted in Dhall files, and hope that this will not cause unintuitive  errors in the future.