[fix] Fix bug with as being a prefix of assert in simple-label
There is a bug in the simple-label rule keyword simple-label-next-char+ because there are keywords that are prefixes of others. This rule would allow "assert" as a simple label since it matches the keyword as followed by the simple-label-next-char list sert. To avoid this, we introduce two new rules that say that as may be a prefix of a valid simple-label only if the simple label is not assert.