[feature] Add very preliminary support for custom selection handling
This adds Dhall-specific extensions for the basicWordSelectionFilter, which disables IntelliJ's automatic basic word selection for Dhall, and extendWordSelectionHandler, which adds the custom selection handling. It is necessary to implement the basicWordSelectionFilter because by IntelliJ's selection rules, the smallest selection wins. The basic word selection mechanism is intent on selecting single characters (largely because of how the grammar is written), which will therefore always be taken as the default iregardless of our custom selection implementation, so we disable it.