[fix] Fix further selectioner behavior and parsing of IPv6 literals
This adds fixes to ensure that record literals and identifiers with builtin or keyword prefixes are highlighted correctly. WAlso fixed is the highlighting of IPv6 literals, which were incorrectly transcribed from the specification. New tests are also aadded for IP literal highlighting. Selectioner behavior is now limited to PsiElements in the Dhall language.