[test] Add more syntax highlighter and selectioner tests
Syntax tests (used by the syntax highlighter and selectioner so far) are refactored to be separated into different traits so that we can have a standardized series of tests for syntax for each feature. Tests in files have been mostly refactored away in favorr of tests in strings inline with the code, which makes it much easier to add a new test.