[package] Create method for releasing and building package on CI
This introduces the sbt release command based off the sbt package. This allows a new version to be released by bumping the version, tagging, committing the change, pushing the change to master (which should trigger the Deploy stage of the CI build, which goes off only on tags), and committing a new bumped snapshot version of the change, which can be pushed to master or a branch if desired.