[package] Prepare for first release
A tricky thing about using the SBT IDEA plugin to patch xml is that it only works on p product files within the SBT settings key productDirectories. To correctly patch the XML,  we use packageArtifact to generate the files into the productDirectory. Then, we run it again to patch the productDirectory XML. Finally, we run packageArtifactZip to zip all the fgenerated files into a ZIP for uploading to the plugin repository. We also add several Dockerfile commands and slightly generalize the Dockerfile so that its entrypoint is just sbt and not specifically sbt test