[fix] Ensure all tagged color settings in the example text are editable in the color settings page
Several tests are added to ensure that all text attribute keys that appear in the tags used in the example text are editable on the text settings page, and that all custom  Dhall syntax highlight attribute keys are in the tags available to the example text. For now, we presume that the example text uses all XML tags available to it, but may test for this in the future.